Aligned some features closer to native Ring behavior
Made it so you can directly switch between armed modes without having to disarm first. Also extended it so you could change mode while it's still arming, although this is not currently supported by Alarmo (It simply won't process the new arm service call).
Also adjusted the countdown event so it uses the next highest countdown feedback, I think this is better than hardcoding the exact amounts, because the final "armed" or "alarm" event will interrupt it anyway, so even though the countdown doesn't quite get to finish if for example set to 10 seconds and using the 15 second countdown, the behavior is still sensible and better than the fallback error beep.
Added panic alarm handling, right now this only sets the alarm to pending, but I am already in touch with the Alarmo devs to get that corrected.